+++ src/CodeNexus.Domain/Entities/Notification.cs
@@ -26,7 +26,7 @@ namespace CodeNexus.Domain.Entities
         public Guid? LessonId { get; set; }
         public Guid? LearningPathId { get; set; }
         public string? NotifiedPathTitle { get; set; }
-        public int? NotifiedSourceVersion { get; set; }
+        public decimal? NotifiedSourceVersion { get; set; }
         public string? NotifiedMentorUserName { get; set; }
         public bool IsRead { get; set; } = false;
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
